Refactor code, separate elasticsearch tasks out of common.py
